fix(nhost): move index.ts to dataProvider folder
